apps/agente-fiscal/requirements.txt
python-dotenv>=1.0.1
requests>=2.31.0
pdfplumber>=0.11.0
